requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
pymongo
setuptools
Certifi
pymongo[srv]==3.12
dotenv
dill 
pyaml



##-e .